* configure.in: Add checks for system headers that are required         before using AC_FUNC_SELECT_ARGTYPES.  Workaround for autoconf         2.13 bug.
I'm only checking in the configure.in and configure script.  When you
build, it will probably rebuild all the Makefiles ...

By the way: I have patched my version of autoconf to fix the bug that
this change works around.  As a result, the checks for some header
files are duplicated.  Nothing to worry about; it will go away as soon
as someone rebuilds configure with an unpatched autoconf 2.13.