Added note about moving extras to modules.
The repository should be stable again.  If not, flames to me.

After doing a "cvs update", you will find that a "make" fails.  This
is because the makefiles still have extras in them.  To fix this, just
do a normal "make", let it fail, then run "config.status" from the top
of the build directory.  This rebuilds the Makefiles, and "make" will
now run.