Ensure that global (i.e. non-static) functions are properly prototyped.
This pass was just the libs/ directory.  Basically, anything that
wasn't already prototyped got marked static unless the function was used
in another file, in which case I added a prototype.  One function
(Parse.c:CheckNTokens) was not used *anywhere*, so I commented it out.
If this is for future use, I apologize ...

Created four new .h files in the process.