* mod_m2f_communication.html: Added description of InitGetConfigInfo.
* TODO (Bugfixes): Added note,  part of the "Make GetConfigLine be
more intelligent" is done.

* FvwmIconMan/readconfig.c (init_config_file):
* FvwmWharf/FvwmWharf.c (ParseOptions): Use InitGetConfigLine to
reduce pipe traffic.

I'm not sure that InitGetConfigLine is  correct for FvwmConsole.  If I
understand the  functionality of FvwmConsole, you could  hook it up to
things that might want the whole module configuration?

I think thats it for InitGetConfigLine.  It doesn't seem to me like
this should be in the NEWS file.