Fix y-position of windows using MinOverlapPlacement.
Due to how the offsets for the best window positioning is calculated, the y
position was not ever taking into account the placement of the window with
respect to the size of the borders; effectively pushing it off-screen.  Fix
it by resetting the y position to take into account the total size of the
border height.