Add new MenuStyle option "[!]UniqueHotkeyActicatesImmediate"
This change allows for the hotkey to be pressed on a menu and for it
remain open until the user presses one of the keys to enact that
specified item.  The default behaviour is the current one of closing
the menu and running the action in the case of there being only one
entry at the specified hotkey.