Add option "automatic" to direction sub-command of Resize.
With "Resize direction automatic", this allows for the resize command to
work out for itself which direction a window should be resized in based on
the pointer's location within a given "quadrant".

Such a quadrant is the same as the 3x3 grid drawn when the "ResizeOutline"
style option is used.