Support XRandR and remove Xinerama
This crudely rips out Xinerama and in its place adds support for XrandR
instead.

Screens are no longer referenced by a number, instead the display name of
the output is needed, as defined by xrandr(1).  The same goes for any
geometry string specifiers in FVWM which might have referenced a screen,
such as:

PositionPlacement 100+200x10+10@0

The "@0" part of that must be changed.

The API is kept intact as much as possible in libs/FScreen.[ch]; with a few
deprecations which were Xinerama-specific and no longer needed.  Previous
screen specifiers used to differentiate at the API level:

FSCREEN_XYPOS
FSCREEN_CURRENT
FSCREEN_PRIMARY
FSCREEN_GLOBAL

Global is now referenced by the output name of "global", and is also the
monitor used in the case where XRandR is not used.  There is no longer the
concept of a primary output (which was only ever used in cases where it was
not known which monitor a window should be placed on.)  CURRENT and XYPOS
now also collapse into the same meaning, but the API hasn't (yet) been
updated to reflect that.

There is also (at present), no support for XRRUpdateConfiguration() on the
root window, hence unplugging and changing the rotation of monitor is not
supported at this time.