Regenerate FVWM/Commands.pm
So it turns out that I didn't do this for the 2.6.0 release; thankfully
there weren't any new commands added then.

Apparently, the way the commands are generated for FVWM/Commands.pm relies
on some *very* carefully formatting of comments in the functable.c file;
which I didn't follow for InfoStore{Add,Remove}.  Hence, garbage produced
until I fixed that.

Ideally this needs revisiting in the future...