Remove debian/ and rpm/ directories
The package-specific mechanisms to build FVWM for deployment on Debian or
RPM-based Linux distributions are best left to the packagers themselves.
Having this is in FVWM isn't helpful to the packagers, and arguably has
nothing to do with window management.

For those developers who don't even have a Linux platform, maintaining these
files is impossible.