Fix disappearing windows.
The old patch that removed synthetic UnmapNotify events on the root window to
suppress windows in HandeMapRequestKeepRaised() being unmapped later.  However,
this also caused problems.  vmplayer windows still diasppeared when returning
from fullscreen mode.

It seems that the real problem was the XUnmapWindow in HandleUnmapNotify.
Whoever thought this was a necessary or a good idea is wrong.  When we get an
UnmapNotify on a client window, the client has already unmapped it itself.
When we get a synthetic UnmapNotify on the root window the client has also
unmapped the window.  There's no need whatsoever to unmap it again, and events
caused by unmapping twice do confuse applications.