Add "J" function actions (late immediate).
This works like "I" but is executed later (once), when the user
starts interacting with the function, e.g. on the first button
release or if the pointer moves (and the function uses "M").  This
can be used to write an "escalating" close function that runs
on the first button release:

  AddToFunc CloseOrDestroy
  + J Close
  + D Destroy

Run "Close" immediately (but not right when the button is pressed)
and follow up with a Destroy in case of a double click.  Compared
to the normal "+ C Close" this avoids the double click delay
before the action is triggered.