Fix updating window and icon titles.
With the introduction of the TitleFormat and IconTitleFormat styles, a change
of the window or icon name could affect both titles.  The existing code did not
reflect this and a change in the icon name might not be visible in the window
title and vice versa.  The patch cleans up and unifies handling of changes of
the window and icon names and fixes this problem.

Also, the said patch simply set the default IconTitleFormat to the same as
TitleFormat, so the icon name would never be used anyway.  This commit replaces
the default IconTitleFormat with "%i" instead.