Outputs: allow for global or per-desk configurations
Rather than assuming that desks and pages are common to all outputs FVWM
has detected via RandR, add an option to make this "global" or
"per-desktop".

Currently, the existing behaviour is to operate in a global mode,
whereby a change of desk or page happens across all outputs.  Sometimes
though, it's desirable to have desks/pages contained to the outputs
their on.

To achieve this, a new command 'DesktopConfiguration' has been created,
taking two options:

* "global" -- not specific to one output, but applies switching
  desks/pages across all monitors
* "per-monitor" -- makes desks/pages specific to each output such that
   switching to a specific page or desks affects just that monitor and
   no other.

This command can be toggled between global and per-monitor while FVWM is
running and will act accordingly.

FvwmPager has been modified and can be told which monitor to interact
with.  A new option, 'Monitor' has been added and takes the name of the
RandR output to react to.