RandR: monitor subtype events
Rather than assuming the main RandR event is ScreenChange, instead look
at the subtype of the event when a RandR event is received.  This allows
us to better understand how to handle output changes, but also tells us
which output has changed -- meaning that events which include rotation
or resolution changes can be identified more easily.

As a result of this change, callers have had to change when #include the
FScreen API.