RandR: track subtypes for outputs
When a given output changes size, or is disabled, ensure FVWM is better
informed so that it can adjust its view of which outputs are available
and what their sizes are.

FVWM uses rectangle calculations to determine if a window should receive
focus.  When monitors come and go, FVWM needs to reinstate its view of
these outputs and their rectangles, based on the screen size at the
time.

It's more complicated with `DesktopConfiguration per-monitor` due to the
width/height of the screen being different from one another,
potentially.

This change is big, and is littered with debug.

Goes a long way to fixing Github Issue #16, #19.