Global monitor: init and reference correct monitor
When dealing with parameter expansion, ensure we use an appropriate
monitor.  In the case of expanding variables such as $[desk.name<n>], it
was assumed that the fw->m pointer could be used.  However, such
variables don't operate in a window context, and hence, fw is NULL, so
use the current monitor instead.

Additionally, when the global screen is in use via DesktopConfiguration,
ensure we update the changed virtual_scr to *ALL* monitors present,
regardless of whether they're active.