RandR: Go back to tracking global width/height
When dealing with multiple monitors, record the width/height as being
united as one screen, and not keep them separate.  When dealing with
multiple pages, the entire screen needs to be used, otherwise the
offsets for those pages "leaks" on to other monitors.

This change means that the GotoPage & GotoDeskAndPage commands have been
updated to accept a screen name, which is called from FvwmPager.  If no
screen name is given then the current monitor is used.

Because of how FVWM maintains its list of monitors, modules can no
longer access FVWM's internal structure for monitor information.
Rather, monitor information is sent to modules via ConfigInfo.  Only
FvwmPager makes use of this at the moment.

This should go some way to fixing #22, if not entirely.