fix broken positions when drawing 3d borders
The original code always skips the top-right point and bottom-right
point when drawing the border. This is almost invisible if the
background colour is similar to hi and sh colour. However, when the
background colour is very different from hi or sh (for example, flat
borders), then borders appear to be corrupted.

It also underflows when a is 1, in y+i+l-a.