RandR: default to 1.5 upwards
When quering for monitors, use XRRGetMonitors() to derive outputs.  This
allows for easier means for detecting displays from previous versions.

Because of this (and because it means multiple monitors can be treated
as one screen, a la Xinerama), we now can no longer assume that desktops
are the same on all screens.

This change divorces desktop creation from being per-monitor, instead
initialising a list of desktops once at startup (and changed via the
DesktopName command), which is then assigned to all monitors.  This
change allows for greater flexibility when DesktopConfiguration is
per-monitor.  We no longer have to dance through numerous hoops.