StartsOnScreen: honour expanded variables
With FVWM3 being per-monitor aware, there is now a need to be able to
specify the current screen.  However, since that is dynamic, we need to
ensure that the expansion of the current screen happens just in time.

For example:

    Style foo StartsOnScreen $[ponter.screen]

won't work, because when FVWM reads in that line from its config, it
will have expanded $[pointer.screen] to be whatever value is correct at
that time, which means any window named "foo" would start on an
incorrect screen.

Therefore, support expansion of variables at the point we need them for
StartOnScreen.

To achieve this, the following line would be needed:

    Style foo StartsOnScreen $$[pointer.screen]

Which would mean that on the first pass, FVWM would have used:

    Style foo StartsOnScreen $[pointer.screen]

... and then when the value for 'StartsOnScreen' is used, it will
expand $[pointer.screen] to the name of a monitor.