CMD_DesktopName: split into helper functions
Prior to this change, m->DesktopsInfo used a shared pointer to perform a
shallow copy of a reference monitor's desktop information.  This was
primarily used to copy over the names of desktops without the need to
reparse.

However, this will break the ability to set per-screen EWMHBaseStrut
settings since now, the pointer is the same across all monitors, and
they shouldn't be.

So split out the parsing of CMD_DesktopName in to helper functions to
make this easier to apply independantly across monitors.