unmaximize: use window's current screen for positioning
When moving a maximized window between pages or screens, FVWM records
the previous geometry of the unmaximized window, so it can restore
correctly.

However, this doesn't work well when the maximized window itself has
moved between screens or different pages, since the unmaximized geometry
from the point of view of the maximized window is now incorrect.
Subsequently, unmaximizing a window will put that window in to the
ether.

For now, the approach is to make unmaximizing a window imply
MoveToScreen (relative to the window's screen) such that it remains
visible, and should largely do what the user expects.

Fixes #68