FvwmPager: bound windows to monitor viewport
When FvwmPager is in "per-monitor" mode (that is, been configured with
"*FvwmPager: Monitor <name>"), it is only ever useful to show the
windows which are on that monitor.

To achieve this, we therfore need to translate the windows's coordinates
to the bounding "view" the monitor the window is on has (that is, its
own width/height).

Translate the window's position (which is relative to the entire XServer
display) to the appropriate monitor, and keep tracking where it is.

This rewrites the CalcGeom() function which previously used a "mirroring"
technique to reflect the bounds of the window in the pager.

Helps #44