UPDATE_FVWM_SCREEN: respect StartsOnDesk style
Previously, when StartsOnDesk was used, although the desk was correctly
set, the desk would never change thereafter, once the window was moved.

This was because the UPDATE_FVWM_SCREEN macro assumed it could always
update the window, but never took into account that when moving
monitors, the desk also needed to be changed.

But this can't blindly happen in all cases, as on initla mapping of the
window, the UPDATE_FVWM_SCREEN macro would have overriden the
StartsOnDesk style.

This change introduces a flag to track this as well as reducing the
number of assigments which weren't needed when updating the screen.
Also removed an extraneous assignment.

Helps #86