configure: resurrect VERSIONINFO
When embedding the git revision string into the version number of FVWM3,
don't overwrite VERSION with the git SHA1.

VERSION is required as the released version so that modules are
maintained in a correct manner, and that version checks for things like
state files (across FVWM3 reboots) are still honoured correctly.