Capture: force NormalState despite current desk
When FVWM restarts and captures windows, the window would only have been put
in the NormalState if the window's desk was different to the current desk.

This check doesn't need to happen any more, since the desks will likely almost
always be different when capturing windows (especially if per-monitor mode is
in use).

Fixes #97