session: fix version check for used_sm
When FVWM3 restarts and then tries to recapture windows, it will first
check that the version of FVWM3 restarting is the same as the version
recorded in the state file.

Prior to this change, there was an erroneous comma, due to the
reproducible build patch, causing this check to fail.