restart: LoadwindowStates: fix NULL dereference
When restarting fvwm3, a state file is written to disk to record window
states/positions so that when they're recaptured by fvwm3, the windows
are as they were before.

The session code assuems that there is at least one window, which isn't
necessarily true.  This then throws out the reference counting when
reading in information which isn't based on a window -- such as monitor
information.  In such cases, fvwm3 would segfault.

This is changed to resize the matches list.