setup_window_placement: initialise screen correctly
When setting up the window, set the screen flags for the specific
screen, if StartsOnScreen is used.  The screen is only set up once here,
before the information is set to the placement algorithms.

Prior to this change, the screen wasn't being expanded, and was always
returning a blank screen.  Furthermore, this was throwing off other
calculations.

Fixes #115