monitor: track focus event separately
When modules learned of the ability to send monitor information, this
was previously sent on all window focus events, as the FVWM modules
required this to reset which monitor had the focus.

This is not ideal with other modules which end up consuming this
information when they might not care about monitor information at all.

Instead, remove BroadcastMonitor() from FocusIn events and set FvwmPager
to be a listener of MX_MONITOR_FOCUS.