Github Actions: switch to using Docker
Many of the free-to-use CI systems work great, but lack environments
which are up to date.  Github Actions is no exception, alas.

In order to ensure that newer version of libraries can be accessed, use
a Docker environment.  The one Github Actions uses is stored on
fvwmorg's Package site, which has pre-existing libraries installed.

Any updates to dependencies will mean that the fvwm3-build:latest docker
environment will need updating.