FvwmForm: resurrect from retirement
A few years ago, in conversation with various people on IRC and some via
email, it was clear then that FvwmForm could be deprecated in favour of
FvwmScript because that was the module which was being more widely used,
and had support for widgets which FvwmForm did not.

However, in speaking to Dan Espen (the primary author of FvwmForm), he
always had the idea to augment FvwmForm with better features than
FvwmScipt.  Indeed, Dan makes a good point that FvwmScript's comments
are in French which doesn't make it easier to non-French speakers to
understand them, *and* follow the code at the same time.

The default assumption in deprecating FvwmForm for FvwmScript -- because
FvwmScript was deemed a superset -- doesn't hold try any more.  If
anything, it's more likely to go the other way; removing FvwmScript for
FvwmForm.

Until we work out what do to about this, we'll need to reinstate
FvwmForm in order to start using it.

There's a few tweaks to the code to make it compile against newer APIs
introduced since FvwmForm was deprecated.

Thanks to Dan Espen for his patience!