fvwm-menu-headlines: retire
fvwm-menu-headlines was originally created to allow the web to be
scraped from various news sites to produce a menu of entries.  A great
idea, but with the advent of RSS feeds and other aggregation programs,
this script has languished over the years.

The sites in question no longer function in the way the script assumes,
and keeping this up to date with largely dynamic HTML and AJAX requests,
makes this script's job almost impossible to do.

Goodbye to it.