Introduce FvwmPrompt
This introduces a new command, FvwmPrompt, as a replacement for
FvwmConsole and FvwmCommand.

As a replacement for FvwmCommand, it can be run as:

    FvwmPrompt 'fvwm command here'

As a replacement for FvwmConsole, it can be run as:

    FvwmPrompt

Unlike FvwmConsole, FvwmPrompt is a standalone program communicating
with FvwmMFL, and is not therefore a Fvwm module.  As such, FvwmPrompt
does not accept the same command-line arguments as FvwmConsole to get it
to launch a specific terminal (-terminal), for example.  Instead, a
compat-friendly shell alias is recommended (but not provided):

   alias FvwmConsole='xterm -g 80x24 -e FvwmPrompt'

This also means that FvwmPrompt does not need to be started from FVWM
either, and will work as long as "Module FvwmMFL" is running.  This is
the case by default, and hence no configuration changes should be
required to make FvwmMFL load.

To enable FvwmPrompt, autotools will look for the Go runtime installed
on the system, and must be at a version >= 1.4.x.  It must also be told
to do this, via --enable-golang, as in:

    ./configure --enable-golang

It is expected in the future that more modules could be written in Go,
hence this option.  It is not currently mandatory that Go needs to be
installed, hence this configure flag.

If '--enable-golang' is used, and configure checks pass, then this
explicitly bypasses compiling and installing FvwmConsole.  This is a
compatibility feature only for now, until such time that confidence in
FvwmPrompt's use and surrounding tooling on different platforms (BSDs
mostly) is met without any large overhead for package maintainers.