Env: add FVWM_IS_FVWM3 envvar
Because fvwm3's version checking resets to 1.0.0, this makes things
awkward for people trying to reuse their existing fvwm2 configs,
especially if they're checking for a string such as >=2.6

Since >2.6 won't work to imply fvwm3 (and that this will be removed in
fvwm3 soon anyway), instead provide an environment variable:
FVWM_IS_FVWM3 which can be used with EnvIsSet test condition.

Somewhat addresses #195 but not fully.  See the comments therein.