FvwmPager: respect DesktopConfiguration setting
When FvwmPager is launched without a "Monitor" config, it assumes the
remit is global -- that is, it will show windows from all monitors.
However, this could be at odds with the DesktopConfiguration setting.

When FvwmPager performs tasks such as switching pages and/or desks, it
does so by sending Fvwm commands directly.  This includes monitor
information which is not what's needed if DesktopConfiguration is
global.

Therefore, only send the monitor string across if FvwmPager is already
tracking a specific monitor.

Helps with #236