Introduce 'Status' command
The Status command writes information about all configured screens,
desks, and clients in JSON format down a named pipe (fifo) for
consumption by third-party tools such as polybar, lemonar, etc.

Unlike other modules which provide specific information, piecing
together a set of data for an "overview" is problematic, and would
require a few hoops to jump through.  This needs to be as fast as
possible.

Fixes #253