HandlePaging: ignore bounds check for panframes
When panframes are entered to enable paging (either by dragging a window
over them, or the mouse entering them), there has always been a check to
ensure that the pointer location is within the window, as these
panwindows would often get buried.

However, the check for this has already been done, so it can be removed.
Additionally, removing this check helps to correct handle changing pages
for monitors whose resolutions are different.