FvwmPager: fix geometry string values
When FvwmPager parses a geometry string with a monitor name embedded in
it, such as:

    *FvwmPager: Geometry 100x90-0+0@MONITORNAME

FvwmPager would not take into account MONITORNAME's dimensions, instead
defaulting to the current monitor.

This results from an incorrect use of global screen, vs specific screen.
Geometry strings are treated as the global screen, so FvwmPager needs to
use those values when calculating where to place its window.

Fixes #310