FvwmEvent: remove support for Rplay
librplay support in FvwmEvent stems from when it was originally known as
FvwmAudio to play sounds on specific events.

This was a popular trend in the 90s, but has become rather annoying in
recent years for most users.  Indeed, even librplay itself is not
maintained and the cost of some Linux distributions carrying it around
just to have the feature enabled in fvwm, can be annoying, despite it
being an optional dependency.

Therefore, remove librplay support.

Fixes #316