libs: remove CatString2/CatString3
CatString2 and CatString3 provided a means of concatenating either two
or three variables together.  However, the callers of these don't get to
specify a format string, as in:

    "%s %s"

which breaks compiling fvwm3 with:

    CFLAGS="-Wformat -Werror=format-security"

Given these wrappers are inflexible, and that xasprintf() already exists
in the codebase, switch to using that.

Fixes #334