FvwmPager: don't update desk directly
When a window is being moved within the pager, the desk the window is on
was being updated, and then later on, FVWM_UPDATE_SCREEN would further
confuse things.  FvwmPager sends Move/MoveToDesk commands back to Fvwm
to get the window's properties updated.

Let the FVWM_UPDATE_SCREEN logic update the window, since the Move
command FvwmPager sends back will also update the desk the window is on,
so no need to explicitly move the window.

Helps #198