desktopconfiguration: tigten global mode check
When issuing a `DesktopConfiguation global` command which is already
using the global mode, the end result is a no-op, yet the code will
still iterate over all monitors looking for something to do.

Correct this by only switching to global mode if `DesktopConfiguation`
is set to something other than `global`.