FvwmMFL: make pid file slightly more unique
When FvwmMFL learned to only start one instance of itself, the pid file
used didn't contain any unique attribute.  As such, trying to run fvwm3
with FvwmMFL under Xephyr meant that separate instance of FvwmMFL was
never started.

To fix this, embed the DISPLAY value to the pid filename to allow it to
run more than once, albeit unique to each display.