update_fvwm_screen: don't lose desk from StartsOnScreen
When updating the desk which a window is on, don't be so clever about
updating the currentdesk if it's the same as the screen.  Other parts of
fvwm3 will be setting this conditionally for their own reasons, and this
check was clobbering that (such as in placement.c for StartsOnDesk
style).

Fixes #373