remove -s option; always use default screen
Before the days of actual Xinerama or XRandR integration, it was possible to run fvwm on separate displays.
These days though, RandR support is sufficient such that fvwm shouldn't need to start up on separate displays.
Indeed, a downside to this is that it is not possible for windows to move between displays.
Given the maturity of RandR, this mode in fvwm3 should be removed.