fvwm2: provide compat wrapper FvwmCommandS
When fvwm3 deprecated FvwmCommand and replaced it with a wrapper round
FvwmMFL, it no longer required the use of FvwmCommandS.

However, because fvwm3 is currently backwards compatible with fvwm2, the
config files are mostly the same.  To help with reducing the need for
changes, provide the wrapper FvwmCommandS for starting FvwmMFL instead.
Under fvwm2, this will use the older FvwmCommandS code.

Fixes #391