API: introduce API to resolve symbolic monitors
Monitors can already be referred to via geometry strings in the form:

    100x100+0-0@FOO

... where "FOO" can be an actual RandR monitor name, or a symbolic one,
as in:

  * g == global
  * p == primary
  * c == current

Expand this concept to allow these symbolic names when looking up
monitors in general.  This allows modules flexibility without having to
understand many of the internals of how fvwm structures monitors.

This change therefore makes `monitor_by_name` static to FScreen.c, and
introduces the replacement of `monitor_resolve_name` which handles both
a literal RandR name as well as symbolic ones.

Fixes #396