vpacket: augment CONFIGARGS with monitor_name
When Fvwm modules request a list of windows (or windows are sent to
modules via ConfigureNotify), the monitor name was never sent along,
which meant there were a few extra hoops needed to ascertain which
monitor a window is on.

Rather than forcing the user to resolve this, augment CONFIGARGS to
include the monitor name the window is on.  Currently only used by
FvwmIconMan, but this would help with FvwmPager in the future.

Fixes #396