FvwmIconMan: handle 'resoltuion screen <x>'
FvwmIconMan has always made an effort to track windows on different
screens.  With the introduction of RandR, as well as
`DesktopConfiguration per-monitor`, this broke how FvwmIconMan saw
different resolutions, especially for `screen` and `desk`.

This teaches FvwmIconMan about how to resolve its 'screen' resolution
better.  By default, the line:

    *FvwmIconMan: resolution screen

takes into account the *primary* screen.  If might not be the same as
the screen that FvwmIconMan is on, in which case the following can be
used:

    *FvwmIconMan: resolution screen c

The following two lines are equivalent:

    *FvwmIconMan: resolution screen g
    *FvwmIconMan: resolution global

Fixes #396