Avoid hard coding of version info in configure.ac Instead, the shell program, fvwm-version-str.sh will use git describe and git show to obtain tag and release information from the log.
If the annotated release tag is at the HEAD, the version will be
considered a release. If there are commits after the release tag,
the micro version will have the tag offset as well as the commit
of the HEAD appended to it. Going forward, versions will appear as:

Fvwm3 Configuration:

  Version:     1.0.3 (released)

  Executables: /usr/local/bin
  Man pages:   /usr/local/share/man
  Modules:     /usr/local/libexec/fvwm3/1.0.3

for a Release, and

Fvwm3 Configuration:

  Version:     1.0.2-25-b9435cd (not released)

  Executables: /usr/local/bin
  Man pages:   /usr/local/share/man
  Modules:     /usr/local/libexec/fvwm3/1.0.2-25-b9435cd

when HEAD commit <> release tag commit. This example indicated there
are 25 commits after the last release.

The program fvwm-version-str.sh can be run independently to test
the impact of different scenarios. It is fuly documented and may be
altered to taste.

The file, configure.ac currently searches for annotated tags, not
lightweight tags. Removing the option -r from the m4_esyscmd_s command
will allow it to search for the latest tag.

m4_define([v_maj],		[m4_esyscmd_s([./utils/fvwm-version-str.sh major -r])])

On branch ph/gitdesc
	modified:   Makefile.am
	modified:   configure.ac
	modified:   utils/fvwm-version-str.sh