Fix invalid read of size 1 error from Valgrind.
After __handle_bpress_action is called, the window fw might have been
freed by CMD_close. Check that the window still exists before checking
if it is scheduled to be raised, so that we don't attempt to read freed
memory.

fw has been freed but not set to NULL, so it's a dangling pointer.
destroy_window() seems to remove fw from the window list at
addwindow.c:3367. If Scr.FvwmRoot in check_if_fvwm_window_exists is part
of the same linked list then fw won't be in the list so the condition at
line 1727 of events.c will short-circuit and not check
IS_SCHEDULED_FOR_RAISE(fw) thereby avoiding the use-after-free.

Fixes #418