PanFrames: correct right panframe location
When changing EdgeThickness, e.g. from default (2) to 1, PanFrameRight.win is
reconfigured with the wrong position (lower right corner instead of upper
right corner). Moving to a page at the right of the current page with the
mouse does not work as expected after that.

From Göran Bengtson on fvwm-workers@

Fixes #436