PanFrames: handle DesktopConfiguration changes
When panframes are created, the DesktopConfiguration is taken in to
account.  This means that during startup, the panframes will be
initialised in the global state, but could change to being per-monitor
shortly afterward.

The implementation assumed that all monitors would needed to have been
checked, however this isn't true in the case of per-monitor mode; as a
result, panframes around other monitors were being unmapped.

Instead, callers of checkPanFrames() already know which monitor needs to
be involved and can pass that through.  This change removes the previous
should_free_panframe() call as it's never used from more than one place,
so can be inlined once more.  In doing so, s/Bool/bool/ since Bool is
reserved for X11 purposes.

Fixes #381, fixes #401