PanFrames: update global state consistently
When paging is initiated through the use of panframes, they are mapped
and unmapped accordingly.  When `DesktopConfiguration` is in `global`
mode (the default), the global state aggregated across all monitors is
updated.

This update didn't include the state of panwindows, which meant that
once someone had caused paging to happen enough time, panframes which
should have been mapped weren't.

Fixes #459