API: remove MyDisplay{Width,Height}
When Fvwm was not per-monitor aware, it used to maintain the overall
width and height of all attached monitors in a global struct.  These
values were passed around and used as normal.

When Fvwm3 gained the ability to have per-monitor desktops, the screens'
overall width and height became less useful, as the individual screen is
what was required, and the usefulness of storing a global state for all
widths and heights across monitors isn't useful.

Instead, don't store the global width/height for all monitors
per-monitor, and instead have a set of global functions to return these
values instead.

The coccinelle script used to perform this translation is part of this
commit.