Fix potential buffer overflow in session handling code.
This buffer overflow is fixed by including field width limits in the
format strings passed to sscanf.

sscanf's %s and %[ format specifiers can overflow the output buffer unless a
field width limit is given. if sscanf encounters a string without
spaces (with %s) or without the characters in the set given (with %[) and this
string is longer than the output buffer, it will overflow the buffer.