EWMH: SetDesktopNames should set num of desks
FVWM was one of the first WMs to adopt the EWMH specification alongside
the ICCCM2 specification.  As such, some of the means of specifying
things like number of desktops can come about in a few different ways.

The `DesktopName` command sets the names of each desktop (and therefore
implicitly how many desktops there are), but there's also a
`EwmhNumberOfDesktops` command to do the same thing.

Regardless of which command a user choses, `DesktopName` should set the
NumberOfDesktops variable which calculates how many desktops there
actually are.

Fixes #477