UPDATE_FVWM_SCREEN: update EWMH desktop
When a window is dragged between monitors, update the EWMH current desk
if it differs, so that the true current desk is updated.  This has the
(correct, IMO) consequence that the current EWMH desktop changes
depending on which screena command such as

`xprop -root -spy | rg _NET_CURRENT`

is run from.  However, in the case of `DesktopConfiguration` being set
to `shared` or `per-monitor`, this should always be the case, as each
monitor is allowed to be on different desktops.

Fixes #477