Add a screen option to the Scroll command.
This extends the Scroll command by adding an 'screen RANDRNAME' option
to specify which monitor to scroll. This is useful when using per-monitor
or shared DesktopConfiguration and wanting to scroll a monitor other
than the current monitor. Mostly used by FvwmPager when tracking a single
monitor to scroll the correct monitor no matter which screen the pager is on.